don't increment version before destroy
If an embedded document has its version incremented, but an ancester
is flagged for destroy and the root document is not, then this causes
conflicts, which are written to the otherwise empty embedded document,
leaving a corrupt document in storage.

Note that this only happends with the parent document cascades
callbacks.