fix problem where last seems to cache
After making the change to logic for incrementing the version number
this spec started failing even though the behavior was actually correct.
Though post.history_tracks returned an updated list of records, the
call to #last continued to return the same "destroy" track, causing the
comment to be created twice, instead of being recreated and then
redestroyed. Addressing the last record by index appears to have
resolved the problem.